REFACTOR:Folder structure
