+++ src/modules/landing-page/components/event-cards.component.tsx
