+++ src/modules/landing-page/components/events-section.component.tsx
