+++ src/modules/landing-page/components/main-event.component.tsx
